Move menu utils to its own file
